feat: xhr interceptor
