+++ src/errors/ajax.ts
@@ -25,30 +25,6 @@ class AjaxErrors extends Base {
     if (!window.XMLHttpRequest) {
       return;
     }
-    const originalXHR = window.XMLHttpRequest;
-    function ajaxEventTrigger(event: any) {
-      const ajaxEvent = new CustomEvent(event, { detail: this });
-      window.dispatchEvent(ajaxEvent);
-    }
-    function customizedXHR() {
-      const liveXHR = new originalXHR();
-
-      liveXHR.addEventListener(
-        'readystatechange',
-        function () {
-          ajaxEventTrigger.call(this, 'xhrReadyStateChange');
-        },
-        false,
-      );
-
-      return liveXHR;
-    }
-    (window as any).XMLHttpRequest = customizedXHR;
-    window.addEventListener('xhrReadyStateChange', (e: any) => {
-      if (e.detail.readyState === 1) {
-        e.detail.setRequestHeader('test', 'test');
-      }
-    });
     const xhrSend = XMLHttpRequest.prototype.send;
     const xhrEvent = (event: any) => {
       try {
